去除编译告警
git-svn-id: http://192.168.20.11/sequoiadb/trunk@21374 6121f079-4a18-41d3-8c30-8870a322f2c2

+++ SequoiaDB/engine/include/rtnInternalSorting.hpp
@@ -68,7 +68,7 @@ namespace engine
 
       BOOLEAN more() const
       {
-         if ( _limit > 0 && _fetched >= _limit )
+         if ( _limit > 0 && _fetched >= (UINT64)_limit )
          {
             return FALSE ;
          }
